⭐[update] 优化工具面板
