+++ Editor/Windows/Interface/IAssetRuleAddress/RuleAddress.cs
@@ -6,7 +6,10 @@
 
 using System;
 using System.IO;
+using UnityEditor;
+using UnityEngine;
 using IAddressRule = AIO.UEditor.IAssetRuleAddress;
+using Object = UnityEngine.Object;
 
 namespace AIO.UEditor
 {
@@ -14,6 +17,8 @@ namespace AIO.UEditor
     {
         public class AddressRuleGroupRelative : IAddressRule
         {
+            public bool AllowThread => true;
+
             public string DisplayAddressName => "Location = Group + Relative File Path";
 
             string IAddressRule.GetAssetAddress(AssetRuleData data)
@@ -27,6 +32,8 @@ namespace AIO.UEditor
 
         public class AddressRuleGroupFileName : IAddressRule
         {
+            public bool AllowThread => true;
+
             public string DisplayAddressName => "Location = Group + FileName";
 
             string IAddressRule.GetAssetAddress(AssetRuleData data)
@@ -37,6 +44,8 @@ namespace AIO.UEditor
 
         public class AddressRuleFileName : IAddressRule
         {
+            public bool AllowThread => true;
+
             public string DisplayAddressName => "Location = FileName";
 
             string IAddressRule.GetAssetAddress(AssetRuleData data)
@@ -47,24 +56,21 @@ namespace AIO.UEditor
 
         public class AddressRuleInstanceID : IAddressRule
         {
+            public bool AllowThread => false;
+
             public string DisplayAddressName => "Location = InstanceID";
 
             string IAddressRule.GetAssetAddress(AssetRuleData data)
             {
-                try
-                {
-                    var obj = UnityEditor.AssetDatabase.LoadAssetAtPath<UnityEngine.Object>(data.AssetPath);
-                    return obj != null ? obj.GetInstanceID().ToString() : Path.GetFileName(data.AssetPath);
-                }
-                catch (Exception e)
-                {
-                    return Path.GetFileName(data.AssetPath);
-                }
+                var obj = AssetDatabase.LoadAssetAtPath<Object>(string.Concat(data.AssetPath, '.', data.Extension));
+                return obj.GetInstanceID().ToString();
             }
         }
 
         public class AddressRuleCollectRelative : IAddressRule
         {
+            public bool AllowThread => true;
+
             public string DisplayAddressName => "Location = Collect + Relative File Path";
 
             string IAddressRule.GetAssetAddress(AssetRuleData data)
@@ -75,6 +81,8 @@ namespace AIO.UEditor
 
         public class AddressRuleCollectFileName : IAddressRule
         {
+            public bool AllowThread => true;
+
             public string DisplayAddressName => "Location = Collect + FileName";
 
             string IAddressRule.GetAssetAddress(AssetRuleData data)
@@ -85,6 +93,8 @@ namespace AIO.UEditor
 
         public class AddressRuleRootRelative : IAddressRule
         {
+            public bool AllowThread => true;
+
             public string DisplayAddressName => "Location = Root Relative";
 
             string IAddressRule.GetAssetAddress(AssetRuleData data)
@@ -95,6 +105,8 @@ namespace AIO.UEditor
 
         public class AddressRuleUserDataRelative : IAddressRule
         {
+            public bool AllowThread => true;
+
             public string DisplayAddressName => "Location = UserData + Relative File Path";
 
             string IAddressRule.GetAssetAddress(AssetRuleData data)
@@ -108,6 +120,8 @@ namespace AIO.UEditor
 
         public class AddressRuleUserDataFileName : IAddressRule
         {
+            public bool AllowThread => true;
+
             public string DisplayAddressName => "Location = UserData + FileName";
 
             string IAddressRule.GetAssetAddress(AssetRuleData data)
