+++ Editor/Windows/Interface/IAssetRuleFilter.cs
@@ -8,8 +8,11 @@ namespace AIO.UEditor
 {
     public interface IAssetRuleFilter
     {
+        /// <summary>
+        /// 显示的过滤器名称
+        /// </summary>
         string DisplayFilterName { get; }
-        
+
         /// <summary>
         /// 验证资源是否符合规则
         /// </summary>
