+++ Editor/Windows/Mode/AssetCollectWindow.OnDrawFirstPackageMode.cs
@@ -183,6 +183,9 @@ namespace AIO.UEditor
 
         private void OnDrawFirstPackageMode()
         {
+            GUI.FocusControl(null);
+            LookModeCurrentSelectAsset = null;
+
             if (!Config.EnableSequenceRecord)
             {
                 var content = new GUIContent("请启用序列记录功能");
