+++ Editor/Windows/Mode/BuildMode/AssetCollectWindow.OnDrawBuildMode.FTP.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 89c58b28da204bdfa417795877d04c38
+timeCreated: 1704230133
