+++ Editor/Windows/Mode/EditorMode/AssetCollectWindow.OnDrawItem.cs.meta
